refactor: extract transaction-tag relationship code into dedicated module
Extract transaction-tag junction table operations from tag.rs into new
transaction_tag.rs module to eliminate tight coupling.
